PCI driver added

+++ kernel/include/port.h
@@ -9,6 +9,6 @@ VOID OutWord(UINT16 portIndex, UINT16 value);
 VOID OutLong(UINT16 portIndex, UINT32 value);
 UINT8 InByte(UINT16 portIndex);
 UINT16 InWord(UINT16 portIndex);
-UINT16 InLong(UINT16 portIndex);
+UINT32 InLong(UINT16 portIndex);
 
 #endif
